Conflicts:
	Source/WebCore/bindings/scripts/preprocessor.pm
	Source/WebCore/platform/graphics/GraphicsContext.h
	Source/WebCore/platform/graphics/Path.h
	Source/WebCore/platform/graphics/Pattern.h
	Source/WebCore/platform/network/curl/ResourceHandleManager.cpp

+++ Source/WebKit2/UIProcess/Launcher/gtk/ProcessLauncherGtk.cpp
@@ -35,6 +35,7 @@
 #include <WebCore/ResourceHandle.h>
 #include <WebCore/RunLoop.h>
 #include <errno.h>
+#include <glib.h>
 #include <locale.h>
 #include <wtf/text/CString.h>
 #include <wtf/text/WTFString.h>
@@ -105,11 +106,17 @@ void ProcessLauncher::launchProcess()
 }
 
 void ProcessLauncher::terminateProcess()
-{   
+{
+    if (m_isLaunching) {
+        invalidate();
+        return;
+    }
+
     if (!m_processIdentifier)
         return;
 
     kill(m_processIdentifier, SIGKILL);
+    m_processIdentifier = 0;
 }
 
 void ProcessLauncher::platformInvalidate()
